% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcurR-package.R
\docType{package}
\name{calcurR-package}
\alias{calcurR}
\alias{calcurR-package}
\title{calcurR: Miscellaneous Functions for CalCurr}
\description{
Misellaneous functions used in processing aerial imagery and associated metadata from crewed and uncrewed systems.
}
\keyword{internal}
