% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_aerial_merge.R
\name{check_aerial_merge}
\alias{check_aerial_merge}
\title{Title}
\usage{
check_aerial_merge(
  df,
  merge_tim_delta = 2,
  seq_tim_delta = 2,
  seq_dist_delta = 200
)
}
\arguments{
\item{seq_dist_delta}{}
}
\description{
Title
}
